+++ CHANGELOG.md
@@ -5,9 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## Unreleased
 
 - README updates
-- refactored token & chunk methods for better separation
-- chunk methods work at block level
+- changed token & chunk methods for better separation
+- changed chunk methods to work at block level
 - added bos and eos attributes to Chunk and Token to hold character indices
+- added token method to generate tagset from a set of labels and scheme string
 
 ## 0.1.2